pyproject.toml
[project]
name = "cditvallejo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "geopandas>=1.1.2",
    "inegipy>=2.0.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "osmnx>=2.0.7",
    "scipy>=1.15.3",
]
